[ISSUE#13322] Merge the A2aMaintainerService and AiMaintainerService interfaces.
